Mapper file is created and code update
