+++ src/main/java/com/education/schoolautomation/controller/ClassRoomController.java
@@ -1,34 +1,32 @@
 package com.education.schoolautomation.controller;
 
-import com.education.schoolautomation.dto.ClassRoomDto;
-import com.education.schoolautomation.dto.SchoolDto;
+import com.education.schoolautomation.mapper.ClassRoomMapper;
 import com.education.schoolautomation.request.ClassRoomRequest;
 import com.education.schoolautomation.response.ClassRoomResponse;
 import com.education.schoolautomation.service.ClassRoomService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/class-rooms")
-
+@RequiredArgsConstructor
 public class ClassRoomController {
 
-
     private final ClassRoomService service;
-
-    public ClassRoomController(ClassRoomService service) {
-        this.service = service;
-    }
+    private final ClassRoomMapper classRoomMapper;
 
     @PostMapping
     public ClassRoomResponse create(@RequestBody ClassRoomRequest request) {
-        return toResponse(service.create(toDto(request)));
+        return classRoomMapper.toResponse(service.create(classRoomMapper.toDto(request)));
+    }
+
+    @PutMapping
+    public ClassRoomResponse update(@RequestParam(value = "classRoomId") UUID classRoomId, @RequestBody ClassRoomRequest request) {
+        return classRoomMapper.toResponse(service.update(classRoomId, classRoomMapper.toDto(request)));
     }
 
     @DeleteMapping
@@ -36,43 +34,9 @@ public class ClassRoomController {
         service.delete(classRoomId);
     }
 
-    @PutMapping
-    public ClassRoomResponse update(@RequestParam(value = "classRoomId") UUID classRoomId, @RequestBody ClassRoomRequest request) {
-        return toResponse(service.update(classRoomId, toDto(request)));
-    }
-
     @GetMapping
     public List<ClassRoomResponse> getAll() {
-        return toResponseList(service.getAll());
+        return classRoomMapper.toResponseList(service.getAll());
     }
 
-    public List<ClassRoomResponse> toResponseList(List<ClassRoomDto> dtoList) {
-        return dtoList.stream()
-                .map(this::toResponse)
-                .collect(Collectors.toList());
-    }
-
-    private ClassRoomResponse toResponse(ClassRoomDto dto) {
-        ClassRoomResponse response = new ClassRoomResponse();
-        response.setClassRoomId(dto.getClassRoomId());
-        response.setClassRoomName(dto.getClassRoomName());
-        if (dto.getSchool() != null) {
-            response.setSchoolId(dto.getSchool().getSchoolId());
-        }
-        if (dto.getBranches() != null) {
-            response.setBranches(dto.getBranches());
-        }
-
-        return response;
-    }
-
-    private ClassRoomDto toDto(ClassRoomRequest request) {
-        ClassRoomDto dto = new ClassRoomDto();
-        dto.setClassRoomName(request.getClassRoomName());
-
-        if (request.getSchoolId() != null) {
-            dto.setSchool(SchoolDto.builder().schoolId(request.getSchoolId()).build());
-        }
-        return dto;
-    }
 }
