+++ src/main/java/com/education/schoolautomation/repository/ClassRoomRepository.java
@@ -8,5 +8,5 @@ import java.util.UUID;
 public interface ClassRoomRepository extends JpaRepository<ClassRoom, UUID> {
     ClassRoom findByClassRoomId(UUID classRoomId);
     ClassRoom findClassRoomByClassRoomId (UUID classRoomId);
-    void deleteByClassRoomId(UUID classRoomId);
+
 }
