+++ src/main/java/com/education/schoolautomation/repository/ManagerRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.UUID;
 
 public interface ManagerRepository extends JpaRepository<Manager, UUID> {
-    Manager findByIdentityId(UUID identityId);
+    Manager findManagerByIdentityId(UUID identityId);
 }
