+++ src/main/java/com/education/schoolautomation/repository/SchoolRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.UUID;
 
 public interface SchoolRepository extends JpaRepository<School, UUID> {
-    School findBySchoolId(UUID schoolId);
+    School findSchoolBySchoolId(UUID schoolId);
 }
