+++ src/main/java/com/education/schoolautomation/service/impl/AssistantManagerServiceImpl.java
@@ -3,6 +3,8 @@ package com.education.schoolautomation.service.impl;
 import com.education.schoolautomation.dto.AssistantManagerDto;
 import com.education.schoolautomation.entity.AssistantManager;
 import com.education.schoolautomation.exception.RecordNotFoundExceptions;
+import com.education.schoolautomation.mapper.AssistantManagerMapper;
+import com.education.schoolautomation.mapper.SchoolMapper;
 import com.education.schoolautomation.repository.AssistantManagerRepository;
 import com.education.schoolautomation.service.AssistantManagerService;
 import lombok.RequiredArgsConstructor;
@@ -10,23 +12,22 @@ import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor(onConstructor = @__(@Lazy))
 public class AssistantManagerServiceImpl implements AssistantManagerService {
 
     private final AssistantManagerRepository repository;
-
     private final SchoolServiceImpl schoolService;
+    private final AssistantManagerMapper assistantManagerMapper;
+    private final SchoolMapper schoolMapper;
 
     @Transactional
     @Override
     public AssistantManagerDto create(AssistantManagerDto dto) {
-        return toDto(repository.save(toEntity(dto)));
+        return assistantManagerMapper.toDto(repository.save(assistantManagerMapper.toEntity(dto)));
     }
 
     @Override
@@ -38,79 +39,26 @@ public class AssistantManagerServiceImpl implements AssistantManagerService {
     public AssistantManagerDto update(UUID assistantManagerId, AssistantManagerDto dto) {
         AssistantManager exitAssistantManager = repository.findById(assistantManagerId)
                 .orElseThrow(() -> new RecordNotFoundExceptions(4001, "AssistantManager not found"));
-        exitAssistantManager.setFullName(dto.getFullName());
-        exitAssistantManager.setTckn(dto.getTckn());
-        exitAssistantManager.setAge(dto.getAge());
-        exitAssistantManager.setPhoneNumber(dto.getPhoneNumber());
-        exitAssistantManager.setAddress(dto.getAddress());
-        exitAssistantManager.setSsn(dto.getSsn());
-        exitAssistantManager.setSalary(dto.getSalary());
+
+        AssistantManager assistantManager = assistantManagerMapper.toEntity(dto);
+        assistantManager.setIdentityId(dto.getAssistantManagerId());
         if (dto.getSchool() != null) {
-            exitAssistantManager.setSchool(schoolService.toEntity(dto.getSchool()));
+            exitAssistantManager.setSchool(schoolMapper.toEntity(dto.getSchool()));
         }
 
-        exitAssistantManager = repository.save(exitAssistantManager);
-        return toDto(exitAssistantManager);
+        assistantManager = repository.save(assistantManager);
+        return assistantManagerMapper.toDto(assistantManager);
     }
 
     @Override
     public List<AssistantManagerDto> getAll() {
-        return toDtoList(repository.findAll());
+        return assistantManagerMapper.toDtoList(repository.findAll());
     }
 
 
     public AssistantManagerDto getById(UUID assistantManagerId) {
-        return toDto(repository.findById(assistantManagerId).get());
+        return assistantManagerMapper.toDto(repository.findById(assistantManagerId).get());
     }
 
 
-
-
-    public List<AssistantManagerDto> toDtoList(List<AssistantManager> entityList) {
-        return entityList.stream()
-                .map(this::toDto)
-                .collect(Collectors.toList());
-    }
-
-
-    public List<AssistantManager> toEntityList(List<AssistantManagerDto> dtoList) {
-        //dtoList null ise bana boş bir liste oluşturur
-        if (dtoList == null) {
-            return Collections.emptyList();
-        }
-        return dtoList.stream()
-                .map(this::toEntity)
-                .collect(Collectors.toList());
-    }
-
-
-    public AssistantManagerDto toDto(AssistantManager entity) {
-        AssistantManagerDto dto = new AssistantManagerDto();
-        dto.setAssistantManagerId(entity.getIdentityId());
-        dto.setFullName(entity.getFullName());
-        dto.setTckn(entity.getTckn());
-        dto.setAge(entity.getAge());
-        dto.setPhoneNumber(entity.getPhoneNumber());
-        dto.setAddress(entity.getAddress());
-        dto.setSsn(entity.getSsn());
-        if (entity.getSchool() != null) {
-            dto.setSchool(schoolService.toDto(entity.getSchool()));
-        }
-
-        dto.setSalary(entity.getSalary());
-        return dto;
-    }
-
-    public AssistantManager toEntity(AssistantManagerDto dto) {
-        AssistantManager entity = new AssistantManager();
-        entity.setFullName(dto.getFullName());
-        entity.setTckn(dto.getTckn());
-        entity.setAge(dto.getAge());
-        entity.setPhoneNumber(dto.getPhoneNumber());
-        entity.setAddress(dto.getAddress());
-        entity.setSsn(dto.getSsn());
-        entity.setSalary(dto.getSalary());
-        entity.setSchool(schoolService.findById(dto.getSchool().getSchoolId()));
-        return entity;
-    }
 }
