+++ src/main/java/com/education/schoolautomation/service/impl/ManagerServiceImpl.java
@@ -3,10 +3,10 @@ package com.education.schoolautomation.service.impl;
 import com.education.schoolautomation.dto.ManagerDto;
 import com.education.schoolautomation.entity.Manager;
 import com.education.schoolautomation.exception.RecordNotFoundExceptions;
+import com.education.schoolautomation.mapper.ManagerMapper;
 import com.education.schoolautomation.repository.ManagerRepository;
 import com.education.schoolautomation.service.ManagerService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,11 +17,12 @@ import java.util.UUID;
 public class ManagerServiceImpl implements ManagerService {
 
     private final ManagerRepository repository;
+    private final ManagerMapper managerMapper;
 
     @Override
     @Transactional
     public ManagerDto create(ManagerDto dto) {
-        return toDto(repository.save(toEntity(dto)));
+        return managerMapper.toDto(repository.save(managerMapper.toEntity(dto)));
     }
 
     @Override
@@ -31,52 +32,22 @@ public class ManagerServiceImpl implements ManagerService {
 
     @Override
     public ManagerDto update(UUID managerId, ManagerDto dto) {
-        Manager exitManager = repository.findByIdentityId(managerId);
-        exitManager.setFullName(dto.getFullName());
-        exitManager.setTckn(dto.getTckn());
-        exitManager.setAge(dto.getAge());
-        exitManager.setPhoneNumber(dto.getPhoneNumber());
-        exitManager.setAddress(dto.getAddress());
-        exitManager.setSsn(dto.getSsn());
-        exitManager.setSalary(dto.getSalary());
-        exitManager = repository.save(exitManager);
-        return toDto(exitManager);
+        Manager exitManager = repository.findManagerByIdentityId(managerId);
+        Manager manager = managerMapper.toEntity(dto);
+        manager.setIdentityId(exitManager.getIdentityId());
+        manager = repository.save(manager);
+        return managerMapper.toDto(manager);
     }
-
 
     @Override
     public ManagerDto get(UUID managerId) {
-        return toDto(repository.findById(managerId)
+        return managerMapper.toDto(repository.findById(managerId)
                 .orElseThrow(() -> new RecordNotFoundExceptions(4001, "Manager not found")));
     }
 
-
     public Manager findById(UUID managerId) {
-        return repository.findByIdentityId(managerId);
+        return repository.findManagerByIdentityId(managerId);
     }
 
-    public ManagerDto toDto(Manager entity) {
-        ManagerDto dto = new ManagerDto();
-        dto.setManagerId(entity.getIdentityId());
-        dto.setFullName(entity.getFullName());
-        dto.setTckn(entity.getTckn());
-        dto.setAge(entity.getAge());
-        dto.setPhoneNumber(entity.getPhoneNumber());
-        dto.setAddress(entity.getAddress());
-        dto.setSsn(entity.getSsn());
-        dto.setSalary(entity.getSalary());
-        return dto;
-    }
 
-    public Manager toEntity(ManagerDto dto) {
-        Manager entity = new Manager();
-        entity.setFullName(dto.getFullName());
-        entity.setTckn(dto.getTckn());
-        entity.setAge(dto.getAge());
-        entity.setPhoneNumber(dto.getPhoneNumber());
-        entity.setAddress(dto.getAddress());
-        entity.setSsn(dto.getSsn());
-        entity.setSalary(dto.getSalary());
-        return entity;
-    }
 }
